+++ spc-server/src/util/md.rs
@@ -1,4 +1,0 @@
-//! ## Process Markdown
-//! render Markdown(with latex) to HTML, pre-process math block or inline math
-
-pub use spc_util::md2html;