+++ src/config.rs
@@ -152,7 +152,7 @@ impl Default for SiteConfig {
       domain: "http://127.0.0.1:8080".into(),
       slogan: "Subscription, Publishing and Collaboration".into(),
       verification: "".into(),
-      landing_page: "# Subscription, Publishing and collaboration \n A self-hosted online writing platform which comes as a single executable with feed subscription, publishing writing and live collaboration and many other features. \n Focus on the Markdown content, be it a blog, a knowledge base, a forum or a combination of them. Good fit for individual or small club. \n ## [Explore Here](/explore) \n\n \n ## [Collaborative Writing](/mdpad/) \n ![](https://images.unsplash.com/photo-1675124516926-a0864dea0abd)".into(),
+      landing_page: "# Subscription, Publishing and collaboration \n A self-hosted online writing platform which comes as a single executable with feed subscription, publishing writing and live collaboration and many other features. \n Focus on the Markdown content, be it a blog, a knowledge base, a forum or a combination of them. Good fit for individual or small club. \n ## [Explore Here](/explore) \n\n  ![](https://images.unsplash.com/photo-1675124516926-a0864dea0abd)".into(),
       about_page: "# About".into(),
       my_css: String::new(),
       my_js: String::new(),